Breakout ImageSource into file to resolve class heir/import issues.
